Move test database drivers to Sylius\Tests\Driver
