+++ UPGRADE.md
@@ -39,6 +39,10 @@
 
 * `Sylius\Bundle\CoreBundle\DependencyInjection\Compiler\RoutingRepositoryPass` was removed, implement it yourself.
 
+* `Sylius\Bundle\CoreBundle\Test\MySqlDriver` and `Sylius\Bundle\CoreBundle\Test\PgSqlDriver` were moved to test namespace,
+  provide your own implementation or use Doctrin DBAL's `Doctrine\DBAL\Driver\PDOMySql\Driver` and 
+  `Doctrine\DBAL\Driver\PDOPgSql\Driver` instead.
+
 ### Currency / CurrencyBundle
 
 ### Customer / CustomerBundle
