+++ app/config/config_test.yml
@@ -15,7 +15,7 @@ framework:
 doctrine:
     dbal:
         dbname: "%database_name%_test"
-        driver_class: Sylius\Bundle\CoreBundle\Test\MySqlDriver
+        driver_class: Sylius\Tests\Driver\MySqlDriver
 
 monolog:
     handlers:
